Add attempt with yields
Cannot use main function's yield inside of a child function.
I believe you may be able to make the child function a generator
as well to use yield and then keep running .next() on the child
generator. This will let all lines be ran one by one, but will
add tons of extra steps because of the injected code.

Haven't tried it though, so maybe not.